Change long to int

+++ 100.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 
-int p(long n) {
+int p(int n) {
    if (n == 1) {
        return 1;
    }
@@ -16,13 +16,13 @@ int p(long n) {
 
 int main() {
     int max = 0;
-    long max_i = 0;
-    for (long i = 1; i <= 100000; i++) {
+    int max_i = 0;
+    for (int i = 1; i <= 100000; i++) {
       int current_p = p(i);
       if (current_p > max) {
           max = current_p;
           max_i = i;
       }
     }
-    printf("max = %d, max_i = %ld", max, max_i);
+    printf("max = %d, max_i = %d", max, max_i);
 }
